reusing resolver
