+++ crates/parcel_plugin_resolver/src/lib.rs
@@ -1,5 +1,15 @@
-#![deny(unused_crate_dependencies)]
+#[cfg(feature = "napi_bindings")]
+mod napi;
+
+#[cfg_attr(any, feature = "napi_bindings", feature = "nodejs")]
+pub mod nodejs {
+  pub use super::napi::*;
+}
+
+#[cfg_attr(any, feature = "napi_bindings", feature = "wasm")]
+pub mod wasm {
+  pub use super::napi::*;
+}
 
 mod resolver;
-
 pub use resolver::*;
