+++ crates/parcel_plugin_resolver/src/napi/mod.rs
@@ -1,0 +1,1 @@
+pub mod parcel_v2;
