+++ crates/parcel_plugin_resolver/src/napi/parcel_v2/function_ref.rs
