+++ crates/parcel_plugin_resolver/src/napi/parcel_v2/mod.rs
@@ -1,0 +1,7 @@
+mod file_system;
+mod function_ref;
+mod resolver;
+
+pub use file_system::*;
+pub use function_ref::*;
+pub use resolver::*;
